email page client component
